# .specs/contracts/openapi.yaml
openapi: 3.0.4
info:
  title: Rampa Remittances API
  description: |
    API for Solana-based remittances app targeting Spain/Germany to LATAM corridor.
    
    ## Authentication Flow
    
    This API uses Web3Auth for user authentication with a token exchange pattern:
    
    1. **User Login**: User logs in via Web3Auth (Google, Apple, etc.) on the frontend
    2. **Token Exchange**: Frontend calls `/auth/web3auth/validate` with Web3Auth JWT token
    3. **API Token**: Backend validates Web3Auth token and returns our API JWT token
    4. **API Calls**: Frontend uses our API JWT token for all subsequent business API calls
    
    ## Security
    
    - All business endpoints require our API JWT token (Bearer authentication)
    - Web3Auth endpoints (`/auth/web3auth/*`) are public and don't require authentication
    - Health and inquiry endpoints are public
  version: 1.0.0
  contact:
    name: Rampa Team
    email: support@rampa.com
servers:
  - url: https://api.rampa.com
    description: Production server
  - url: https://staging-api.rampa.com
    description: Staging server
  - url: http://localhost:3001
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /wallet:
    post:
      summary: Create new wallet
      tags: [Wallet]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWalletRequest'
      responses:
        '201':
          description: Wallet created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '400':
          description: Invalid wallet data
        '401':
          description: Unauthorized

    get:
      summary: Get wallet details
      tags: [Wallet]
      responses:
        '200':
          description: Wallet details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '401':
          description: Unauthorized

    put:
      summary: Update wallet
      tags: [Wallet]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWalletRequest'
      responses:
        '200':
          description: Wallet updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '400':
          description: Invalid wallet data
        '401':
          description: Unauthorized

    delete:
      summary: Disconnect wallet
      tags: [Wallet]
      responses:
        '200':
          description: Wallet disconnected successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
  /wallet/balance:
    get:
      summary: Get specific wallet balance
      tags: [Wallet]
      parameters:
        - name: tokenType
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/TokenType'
          description: Token type to get balance for
          example: 'EURC'
      responses:
        '200':
          description: Wallet balance retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  walletId:
                    type: string
                    format: uuid
                  tokenType:
                    $ref: '#/components/schemas/TokenType'
                  balance:
                    type: number
                    format: decimal
        '401':
          description: Unauthorized
        '403':
          description: Profile verification required for this operation
  /wallet/balances:
    get:
      summary: Get all wallet balances
      tags: [Wallet]
      responses:
        '200':
          description: Wallet balances retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletBalanceResponse'
        '401':
          description: Unauthorized
        '403':
          description: Profile verification required for this operation
  /wallet/connect:
    post:
      summary: Connect existing wallet
      tags: [Wallet]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectWalletRequest'
      responses:
        '200':
          description: Wallet connected successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '400':
          description: Invalid wallet data
        '401':
          description: Unauthorized
  /wallet/transfer:
    post:
      summary: Transfer funds
      tags: [Wallet]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '200':
          description: Transfer initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  fromWallet:
                    type: string
                  toAddress:
                    type: string
                  amount:
                    type: number
                  tokenType:
                    $ref: '#/components/schemas/TokenType'
        '400':
          description: Invalid transfer data
        '401':
          description: Unauthorized
        '403':
          description: Profile verification required for this operation
  /wallet/suspend:
    post:
      summary: Suspend wallet
      tags: [Wallet]
      responses:
        '200':
          description: Wallet suspended successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  wallet:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      status:
                        type: string
        '401':
          description: Unauthorized
  /wallet/activate:
    post:
      summary: Activate wallet
      tags: [Wallet]
      responses:
        '200':
          description: Wallet activated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  wallet:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      status:
                        type: string
        '401':
          description: Unauthorized
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Our API JWT token obtained from /auth/web3auth/validate
  schemas:
    TokenType:
      type: string
      enum: [USDC, EURC, SOL]
...
    WalletBalanceResponse:
      type: object
      properties:
        balances:
          type: array
          items:
            $ref: '#/components/schemas/TokenBalance'
    TokenBalance:
      type: object
      properties:
        tokenType:
          $ref: '#/components/schemas/TokenType'
        balance:
          type: number
          format: decimal
        lastUpdated:
          type: string
          format: date-time
    ConnectWalletRequest:
      type: object
      required: [walletType, address, publicKey]
      properties:
        walletType:
          type: string
          enum: [phantom, solflare]
        address:
          type: string
        publicKey:
          type: string
    WalletResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        address:
          type: string
        publicKey:
          type: string
        walletType:
          type: string
          enum: [web3auth_mpc, phantom, solflare]
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        balances:
          type: array
          items:
            $ref: '#/components/schemas/TokenBalance'
    CreateWalletRequest:
      type: object
      required: [address, publicKey, walletType]
      properties:
        address:
          type: string
        publicKey:
          type: string
        walletType:
          type: string
          enum: [web3auth_mpc, phantom, solflare]
    UpdateWalletRequest:
      type: object
      properties:
        address:
          type: string
        publicKey:
          type: string
        walletType:
          type: string
          enum: [web3auth_mpc, phantom, solflare]
    TransferRequest:
      type: object
      required: [toAddress, amount, tokenType]
      properties:
        toAddress:
          type: string
        amount:
          type: number
          format: decimal
          minimum: 0.00000001
        tokenType:
          $ref: '#/components/schemas/TokenType'
        description:
          type: string
...